+++ src/main/java/webserver/path/BasicPath.java
@@ -1,12 +1,12 @@
 package webserver.path;
 
-public enum FilePath {
+public enum BasicPath {
     BASE("src/main/resources/static"),
     HOME("/index.html");
 
     private final String path;
 
-    FilePath(final String path) {
+    BasicPath(final String path) {
         this.path = path;
     }
