+++ src/main/java/webserver/request/Uri.java
@@ -3,7 +3,7 @@ package webserver.request;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webserver.handler.FileHandler;
-import webserver.path.PostRequestPath;
+import webserver.handler.PostHandler;
 
 public class Uri {
     private static final Logger logger = LoggerFactory.getLogger(Uri.class);
@@ -16,7 +16,7 @@ public class Uri {
         validate(uri);
     }
 
-    boolean isSame(final PostRequestPath path) {
+    boolean isSame(final PostHandler.Path path) {
         return path.getPath().equals(uri);
     }
 
@@ -25,7 +25,7 @@ public class Uri {
     }
 
     private void validate(final String uri) {
-        if (!FileHandler.isExistFile(uri) && !PostRequestPath.has(uri)) {
+        if (!FileHandler.isExistFile(uri) && !PostHandler.Path.has(uri)) {
             throw new IllegalArgumentException("404에러");
         }
     }
